requirements.txt
aiogram==3.10.0
soccerdata==1.6.0
pandas==2.2.0
scipy==1.12.0
sqlalchemy==2.0.25
python-dotenv==1.0.1
scikit-learn==1.4.0
joblib==1.3.2
fastapi==0.103.1
pydantic==1.10.12
uvicorn==0.23.2
playwright==1.41.0
playwright-stealth==1.0.6
